New video links generation method (Feature Incomplete)
Using a new "exploit", we can derive the video / clip links from the courseinfo.
From there we can request the video links and download directly from pluralsight, without anything being linked to our account.

However this only works for "semi-public" videos, or "free" content made semi-public by pluralsight.
It is unknown what are the criteria those videos are made public, however.

Word of caution is that pluralsight probably did this for less fortunate people that are unable to pay for the full access. And if we exploit it too much, pluralsight may patch it for good.

This feature can be toggled using the "--new" parameter.